# Speed up date_histogram by precomputing ranges (#61467)